moved examples to src/main; updated mock-server example to use scala-test
